Feat: Move logrus calls into files excluded by wasm
This permits builds of wasm binaries.
