+++ internal/sloghandle/logrus.go
@@ -1,3 +1,6 @@
+//go:build !wasm
+// +build !wasm
+
 // Package sloghandle provides a transition handler for migrating from logrus to slog.
 package sloghandle
