+++ internal/sloghandle/logrus_test.go
@@ -1,3 +1,6 @@
+//go:build !wasm
+// +build !wasm
+
 package sloghandle
 
 import (
